Ajuste de RepositorioTurma
