+++ src/SME.SGP.Dados/Repositorios/RepositorioTurmaConsulta.cs
@@ -725,5 +725,17 @@ namespace SME.SGP.Dados.Repositorios
             }, new { turmaCodigo }, splitOn: "TurmaId, UeId, DreId")).FirstOrDefault();
         }
 
+        public async Task<IEnumerable<long>> ObterTurmasIdsPorUeEAnoLetivo(int anoLetivo, string ueCodigo)
+        {
+            var query = new StringBuilder(@"select t.id 
+                                              from turma t
+                                             inner join ue u on u.id = t.ue_id 
+                                             where t.ano_letivo = @anoLetivo
+                                               and u.ue_id = @ueCodigo
+                                               and not t.historica ");
+
+            return await contexto.Conexao.QueryAsync<long>(query.ToString(), new { anoLetivo, ueCodigo });
+
+        }
     }
 }
