+++ src/SME.SGP.Dominio.Interfaces/Repositorios/IRepositorioTurmaConsulta.cs
@@ -41,5 +41,7 @@ namespace SME.SGP.Dominio.Interfaces
         Task<IEnumerable<GraficoBaseDto>> ObterInformacoesEscolaresTurmasAsync(int anoLetivo, long dreId, long ueId, AnoItinerarioPrograma[] anos, Modalidade modalidade, int? semestre);
         Task<Turma> ObterTurmaCompletaPorCodigo(string turmaCodigo);
         Task<IEnumerable<TurmaDTO>> ObterTurmasInfantilPorAno(int anoLetivo, string ueCodigo);
+
+        Task<IEnumerable<long>> ObterTurmasIdsPorUeEAnoLetivo(int anoLetivo, string ueCodigo);
     }
 }
